Removes the concept of page 'parts'
Context: When WebDKP was first built, there were some ambitious plans
to have it support a drag-and-drop website builder for guilds. This included
a plugin framework of page 'parts', such as html, login, or other widgets
could be placed on the page. This never fully panned out, but there is
still substantial code artifacts from this work. This is a first change in a
series of changes that will start removing this content management
framework.