Migrates from Prototype to jQuery
This commit migrates all remaining uses of Scriptaculous and Prototype
over to jQuery 3+. It involved a migration of all uses of the Builder and
Ajax libraries over to their equivalent in jQuery land.

As part of this change, lightbox also needed to be upgraded because
it had a dependency on Scriptaculous.